Generative AI/Projects/03-Travel_Destination_Recommender/requirements.txt
streamlit
google-generativeai
Pillow
python-dotenv 